Refactored results tools code
